+++ src/clients/sql.ts
@@ -4,40 +4,54 @@ import { NonStorableConfigProfile } from "../context/config";
 import { MaterializeObject } from "../providers/schema";
 import AdminClient from "./admin";
 import CloudClient from "./cloud";
+import * as vscode from 'vscode';
+import { Context, EventType } from "../context";
 
 export default class SqlClient {
     private pool: Promise<Pool>;
     private adminClient: AdminClient;
     private cloudClient: CloudClient;
+    private context: Context;
     private profile: NonStorableConfigProfile;
 
     constructor(
         adminClient: AdminClient,
         cloudClient: CloudClient,
         profile: NonStorableConfigProfile,
+        context: Context,
     ) {
         this.adminClient = adminClient;
         this.cloudClient = cloudClient;
         this.profile = profile;
+        this.context = context;
 
         this.pool = new Promise((res, rej) => {
             const asyncOp = async () => {
-                console.log("[SqlClient]", "Building config.");
-                const config = await this.buildPoolConfig();
-                const pool = new Pool(config);
-                console.log("[SqlClient]", "Connecting pool.");
+                try {
+                    console.log("[SqlClient]", "Building config.");
+                    const config = await this.buildPoolConfig();
+                    const pool = new Pool(config);
+                    console.log("[SqlClient]", "Connecting pool.");
 
-                pool.connect().then(() => {
-                    console.log("[SqlClient]", "Pool successfully connected.");
-                    res(pool);
-                }).catch((err) => {
-                    console.error(err);
-                    rej(err);
-                });
+                    pool.connect().then(() => {
+                        console.log("[SqlClient]", "Pool successfully connected.");
+                        res(pool);
+                    }).catch((err) => {
+                        console.error(err);
+                        rej(err);
+                    });
+                } catch (err) {
+                    console.error("[SqlClient]", "Error creating pool: ", err);
+                    this.context.emit("event", { type: EventType.error, message: err });
+                }
             };
 
             asyncOp();
         });
+    }
+
+    async connectErr() {
+        await this.pool;
     }
 
     /**
@@ -54,7 +68,7 @@ export default class SqlClient {
 
         const schema = this.profile.schema;
         if (schema) {
-            connectionOptions.push(`-csearch_path==${schema}`);
+            connectionOptions.push(`-csearch_path=${schema}`);
         }
 
         return connectionOptions.join(" ");
@@ -92,7 +106,6 @@ export default class SqlClient {
     async* cursorQuery(statement: string): AsyncGenerator<QueryResult> {
         const pool = await this.pool;
         const client = await pool.connect();
-        const id = randomUUID();
 
         try {
             const batchSize = 100; // Number of rows to fetch in each batch