+++ src/providers/auth.ts
@@ -63,8 +63,10 @@ async function loginServer(name: string): Promise<AppPasswordResponse | undefine
 }
 
 interface State {
+    isRemoveProfile: boolean;
     isAddNewProfile: boolean;
     isLoading: boolean;
+    error: undefined | string;
 }
 
 /**
@@ -85,7 +87,9 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
         this.context = context;
         this.state = {
             isAddNewProfile: false,
+            isRemoveProfile: false,
             isLoading: this.context.isLoading(),
+            error: undefined,
         };
 
         // Await for readyness when the extension activates from the outside.
@@ -94,11 +98,24 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
             this.state = {
                 ...this.state,
                 isLoading: false,
+                error: undefined,
             };
         });
 
-        this.context.on("event", ({ type }) => {
+        this.context.on("event", (data) => {
+            const { type } = data;
             switch (type) {
+                case EventType.error: {
+                    const { message } = data;
+                    console.log("[AuthProvider]", "Error detected: ", message, data);
+                    this.state.error = message;
+                    this.state.isLoading = false;
+
+                    if (this._view) {
+                        this._view.webview.html = this._getHtmlForWebview(this._view.webview);
+                    }
+                    break;
+                }
                 case EventType.newProfiles: {
                     console.log("[AuthProvider]", "New profiles available.");
                     if (this._view) {
@@ -129,7 +146,13 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
                     console.log("[AuthProvider]", "New environment available.");
                     if (this._view) {
                         this.state.isLoading = false;
+                        this.state.error = undefined;
 
+                        // Do not refresh the webview if the user is removing or adding a profile.
+                        // The UI will auto update after this action ends.
+                        if (this.state.isRemoveProfile || this.state.isAddNewProfile) {
+                            return;
+                        }
                         console.log("[AuthProvider]", "Triggering configuration webview.");
                         this._view.webview.html = this._getHtmlForWebview(this._view.webview);
                     }
@@ -153,8 +176,14 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
         webviewView: vscode.WebviewView
     ) {
         this.state.isAddNewProfile = false;
+        this.state.error = undefined;
+
         if (appPasswordResponse) {
             const { appPassword, region } = appPasswordResponse;
+
+            // Set the state loading to true. After the new context is loaded
+            // loading will turn false.
+            this.state.isLoading = true;
             this.context.addAndSaveProfile(name, appPassword, region.toString());
         } else {
             // Cancel login process.
@@ -204,7 +233,6 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
                     webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
                     break;
                 }
-
                 case "onAddProfile": {
                     this.state.isAddNewProfile = true;
                     webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
@@ -214,6 +242,35 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
                     const { name } = data;
                     console.log("[AuthProvider]", "onProfileChange(): ", data);
                     this.context.setProfile(name);
+                    break;
+                }
+                // Remove Profile:
+                case "onCancelRemoveProfile": {
+                    this.state.isRemoveProfile = false;
+                    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
+                    break;
+                }
+                case "onContinueRemoveProfile": {
+                    this.state.isRemoveProfile = false;
+
+                    // Set the state loading to true. After the new context is loaded
+                    // loading will turn false.
+                    this.state.isLoading = true;
+                    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
+
+                    const name = this.context.getProfileName();
+
+                    if (name) {
+                        this.context.removeAndSaveProfile(name);
+                    } else {
+                        console.error("[Auth]", "Profile name is not available.");
+                    }
+
+                    break;
+                }
+                case "onRemoveProfile": {
+                    this.state.isRemoveProfile = true;
+                    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
                     break;
                 }
                 case "logInfo": {
@@ -276,11 +333,11 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
 		const nonce = getNonce();
 
         let content = (
-            `<div id="profile-setup-container">
-                <vscode-text-field id="profileNameInput">Profile Name</vscode-text-field>
-            </div>
-            <div id="add-profile-actions-container">
-                <vscode-button id="continueProfileButton" disabled=true>Continue</vscode-button>
+            `
+            <vscode-text-field id="profileNameInput">Profile Name</vscode-text-field>
+            <p id="invalidProfileNameErrorMessage">Profile name must contain only ASCII letters, ASCII digits, underscores, and dashes.</p>
+            <div class="setup-container-actions">
+                <vscode-button appearence="primary" id="continueProfileButton" class="action_button" disabled=true>Continue</vscode-button>
             </div>
             `
         );
@@ -290,13 +347,24 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
         if (profileNames) {
             if (this.state.isAddNewProfile) {
                 content = (
-                    `<div id="profile-setup-container">
-                        <vscode-text-field id="profileNameInput">Profile Name</vscode-text-field>
-                    </div>
-                    <div id="add-profile-actions-container">
-                        <vscode-button id="cancelAddProfile">Cancel</vscode-button>
-                        <vscode-button id="continueProfileButton"
+                    `<vscode-text-field id="profileNameInput">Profile Name</vscode-text-field>
+                    <p id="invalidProfileNameErrorMessage">Profile name must contain only ASCII letters, ASCII digits, underscores, and dashes.</p>
+                    <div class="setup-container-actions">
+                        <vscode-button class="action_button" appearance="secondary" id="cancelAddProfile">Cancel</vscode-button>
+                        <vscode-button class="action_button" id="continueProfileButton"
                         disabled=true>Continue</vscode-button>
+                    </div>`
+                );
+            } else if (this.state.isRemoveProfile) {
+                content = (
+                    `<div>
+                        <p>You are about to remove a profile from your configuration.</p>
+                        <p>Please type <b>${this.context.getProfileName()}</b> to confirm: </p>
+                        <vscode-text-field id="removeProfileNameInput" confirm-data="${this.context.getProfileName()}"></vscode-text-field>
+                        <div class="setup-container-actions">
+                            <vscode-button class='action_button' appearance="secondary" id="cancelRemoveProfileButton">Cancel</vscode-button>
+                            <vscode-button class='action_button' appearance="primary" id="continueRemoveProfileButton" disabled=true>Remove</vscode-button>
+                        </div>
                     </div>`
                 );
             } else {
@@ -304,51 +372,60 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
                 const schema = this.context.getSchema();
                 const cluster = this.context.getCluster();
                 const profileName = this.context.getProfileName();
+                console.log("[Auth]", this.state.error);
 
                 content = (
                     `<div class="profile-container">
-                        <div class="setup-container">
-                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" width="20" height="20" stroke-width="1" stroke="currentColor">
-                                <path stroke-linecap="round" stroke-linejoin="round" d="M17.982 18.725A7.488 7.488 0 0012 15.75a7.488 7.488 0 00-5.982 2.975m11.963 0a9 9 0 10-11.963 0m11.963 0A8.966 8.966 0 0112 21a8.966 8.966 0 01-5.982-2.275M15 9.75a3 3 0 11-6 0 3 3 0 016 0z" />
-                            </svg>
-                            <vscode-dropdown id="profiles">
+                        <!--  The following container is an extract from the guidelines: -->
+                        <!--  https://github.com/microsoft/vscode-webview-ui-toolkit/tree/main/src/dropdown#with-label -->
+                        <div class="dropdown-container">
+                            <label for="profiles">Profile</label>
+                            <vscode-dropdown id="profiles" ${this.state.isLoading ? "disabled=true" :""}>
                                 <vscode-option>${(profileName)}</vscode-option>
                                 ${profileNames.filter(name => name !== profileName).map((name) => `<vscode-option>${name}</vscode-option>`).join('')}
                             </vscode-dropdown>
-
-                            <vscode-button id="addProfileLink">
-                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1" stroke="currentColor">
-                                    <path stroke-linecap="round" stroke-linejoin="round" d="M19 7.5v3m0 0v3m0-3h3m-3 0h-3m-2.25-4.125a3.375 3.375 0 11-6.75 0 3.375 3.375 0 016.75 0zM4 19.235v-.11a6.375 6.375 0 0112.75 0v.109A12.318 12.318 0 0110.374 21c-2.331 0-4.512-.645-6.374-1.766z" />
-                                </svg>
+                        </div>
+                        <div class="setup-container-actions">
+                            <vscode-button class='action_button' appearance="secondary" id="removeProfileButton" aria-label="Remove Profile">
+                                Remove
+                            </vscode-button>
+                            <vscode-button class='action_button' id="addProfileButton" appearance="primary" aria-label="Add Profile">
+                                Add
                             </vscode-button>
                         </div>
                         <vscode-divider></vscode-divider>
-                        ${this.state.isLoading ? `<vscode-progress-ring id="loading-ring"></vscode-progress-ring>` : "<div style=''>Configuration</div>"}
-                        <div class="setup-container ${this.state.isLoading ? "invisible" :""}">
-                            <svg width="20" height="20" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg"><g><path d="M7.99967 1.33313L1.33301 4.66646L7.99967 7.9998L14.6663 4.66646L7.99967 1.33313Z" stroke="currentColor" stroke-width="1" stroke-linecap="round" stroke-linejoin="round"></path><path d="M1.33301 11.3331L7.99967 14.6665L14.6663 11.3331" stroke="currentColor" stroke-width="1" stroke-linecap="round" stroke-linejoin="round"></path><path d="M1.33301 7.99988L7.99967 11.3332L14.6663 7.99988" stroke="currentColor" stroke-width="1" stroke-linecap="round" stroke-linejoin="round"></path></g></svg>
-                            <vscode-dropdown id="clusters">
-                            <vscode-option>${cluster?.name}</vscode-option>
-                                ${(this.context.getClusters() || []).filter(x => x.name !== cluster?.name).map(({name}) => `<vscode-option>${name}</vscode-option>`).join('')}
-                            </vscode-dropdown>
-                        </div>
-                        <div class="setup-container ${this.state.isLoading ? "invisible" :""}">
-                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" width="20" height="20" viewBox="0 0 24 24" stroke="currentColor" stroke-width="1">
-                                <path stroke-linecap="round" stroke-linejoin="round" d="M20.25 6.375c0 2.278-3.694 4.125-8.25 4.125S3.75 8.653 3.75 6.375m16.5 0c0-2.278-3.694-4.125-8.25-4.125S3.75 4.097 3.75 6.375m16.5 0v11.25c0 2.278-3.694 4.125-8.25 4.125s-8.25-1.847-8.25-4.125V6.375m16.5 0v3.75m-16.5-3.75v3.75m16.5 0v3.75C20.25 16.153 16.556 18 12 18s-8.25-1.847-8.25-4.125v-3.75m16.5 0c0 2.278-3.694 4.125-8.25 4.125s-8.25-1.847-8.25-4.125" />
-                            </svg>
-                            <vscode-dropdown id="databases">
-                                <vscode-option>${database && database.name}</vscode-option>
-                                ${(this.context.getDatabases() || []).filter(x => x.name !== database?.name).map(({name}) => `<vscode-option>${name}</vscode-option>`).join('')}
-                            </vscode-dropdown>
-                        </div>
-                        <div class="setup-container ${this.state.isLoading ? "invisible" :""}">
-                            <svg width="20" height="20" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.2" stroke="currentColor">
-                            <path stroke-linecap="round" stroke-linejoin="round" d="M6 6.878V6a2.25 2.25 0 012.25-2.25h7.5A2.25 2.25 0 0118 6v.878m-12 0c.235-.083.487-.128.75-.128h10.5c.263 0 .515.045.75.128m-12 0A2.25 2.25 0 004.5 9v.878m13.5-3A2.25 2.25 0 0119.5 9v.878m0 0a2.246 2.246 0 00-.75-.128H5.25c-.263 0-.515.045-.75.128m15 0A2.25 2.25 0 0121 12v6a2.25 2.25 0 01-2.25 2.25H5.25A2.25 2.25 0 013 18v-6c0-.98.626-1.813 1.5-2.122" />
-                            </svg>
-                            <vscode-dropdown id="schemas">
-                                <vscode-option>${schema && schema.name}</vscode-option>
-                                ${(this.context.getSchemas() || []).filter(x => x.name !== schema?.name).map(({name}) => `<vscode-option>${name}</vscode-option>`).join('')}
-                            </vscode-dropdown>
-                        </div>
+                        ${this.state.error ? `<p class="profileErrorMessage">${this.state.error}</p>`: ""}
+                        ${this.state.isLoading ? `<vscode-progress-ring id="loading-ring"></vscode-progress-ring>` : ""}
+                        ${(!this.state.isLoading && !this.state.error) ? "<span id='options-title'>Connection Options</span>": ""}
+                        ${(!this.state.isLoading && !this.state.error) ? `
+                            <div class="setup-container ${this.state.isLoading ? "invisible" :""}">
+                                <div class="dropdown-container">
+                                    <label for="clusters">Cluster</label>
+                                    <vscode-dropdown id="clusters">
+                                        <vscode-option>${cluster?.name}</vscode-option>
+                                        ${(this.context.getClusters() || []).filter(x => x.name !== cluster?.name).map(({name}) => `<vscode-option>${name}</vscode-option>`).join('')}
+                                    </vscode-dropdown>
+                                </div>
+                            </div>
+                            <div class="setup-container ${this.state.isLoading ? "invisible" :""}">
+                                <div class="dropdown-container">
+                                    <label for="databases">Database</label>
+                                    <vscode-dropdown id="databases">
+                                        <vscode-option>${database && database.name}</vscode-option>
+                                        ${(this.context.getDatabases() || []).filter(x => x.name !== database?.name).map(({name}) => `<vscode-option>${name}</vscode-option>`).join('')}
+                                    </vscode-dropdown>
+                                </div>
+                            </div>
+                            <div class="setup-container ${this.state.isLoading ? "invisible" :""}">
+                                <div class="dropdown-container">
+                                    <label for="schemas">Schema</label>
+                                        <vscode-dropdown id="schemas">
+                                            <vscode-option>${schema && schema.name}</vscode-option>
+                                            ${(this.context.getSchemas() || []).filter(x => x.name !== schema?.name).map(({name}) => `<vscode-option>${name}</vscode-option>`).join('')}
+                                        </vscode-dropdown>
+                                </div>
+                            </div>
+                        `: ""}
                     </div>
                 `);
             }
@@ -372,7 +449,9 @@ export default class AuthProvider implements vscode.WebviewViewProvider {
         </head>
         <body>
             <div id="container">
-                <img id="logo" src="${logoUri}" alt="Materialize Logo" />
+                <div id="logoContainer">
+                    <img id="logo" src="${logoUri}" alt="Materialize Logo" />
+                </div>
                 ${content}
             </div>